requirements.txt
aniso8601==7.0.0
asgiref==3.7.2
Django==3.2.19
django-filter==23.2
django-graphql-auth==0.3.16
django-graphql-jwt==0.3.0
djangorestframework==3.14.0
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
Pillow==9.5.0
promise==2.3
PyJWT==1.7.1
pytz==2023.3
Rx==1.6.3
singledispatch==4.0.0
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.6.3
tzdata==2023.3
